Account for workflows that increment the version prior to publication
Some common workflows increment the version in metadata.json
immediately after publication to represent that current work is towards
the next incremental version release candidate. To compare properly, we
*should* compare the released version with the version of the latest
tag. However, I'm not convinced that we can assume that latest tag
exists consistently. I think it's likely that we'd get far more false
positives than is worth.

Instead, we check to ensure that the Forge version isn't *greater than*
the repository version. This does leave us vulnerable to a small attack
scenario in which the attacker carefully reproduces the same version
recorded in the repository and pushes that to the Forge, but I'm not
sure we can do better than this. We'll have to find other ways to flag
that behaviour.

Fixes #1