Alternate version freshness check
This will compare the Forge release version and the version of the
latest tag. If those are not equal, then it will check to ensure that
the Forge release version is less than or equal to the version in
metadata. I think this ought to be vanishingly close to eliminating the
gap in the previous commit.